# Added negated property to Relational and BooleanAtom