--- a/sympy/core/relational.py
+++ b/sympy/core/relational.py
@@ -124,6 +124,28 @@ class Relational(Boolean, Expr, EvalfMixin):
         a, b = self.args
         return ops.get(self.func, self.func)(b, a, evaluate=False)
 
+    @property
+    def negated(self):
+        """Return the negated relationship.
+
+        Examples
+        ========
+
+        >>> from sympy import Eq
+        >>> from sympy.abc import x
+        >>> Eq(x, 1)
+        Eq(x, 1)
+        >>> _.negated
+        Ne(x, 1)
+        >>> x < 1
+        x < 1
+        >>> _.negated
+        x >= 1
+        """
+        ops = {Eq: Ne, Ge: Lt, Gt: Le, Le: Gt, Lt: Ge, Ne: Eq}
+        a, b = self.args
+        return ops.get(self.func, self.func)(a, b, evaluate=False)
+
     def _eval_evalf(self, prec):
         return self.func(*[s._evalf(prec) for s in self.args])
 
@@ -500,7 +522,7 @@ class Unequality(Relational):
         if evaluate:
             is_equal = Equality(lhs, rhs)
             if isinstance(is_equal, BooleanAtom):
-                return ~is_equal
+                return is_equal.negated
 
         return Relational.__new__(cls, lhs, rhs, **options)
 
@@ -524,7 +546,7 @@ class Unequality(Relational):
         if isinstance(eq, Equality):
             # send back Ne with the new args
             return self.func(*eq.args)
-        return ~eq  # result of Ne is ~Eq
+        return eq.negated  # result of Ne is the negated Eq
 
 Ne = Unequality