--- a/sympy/logic/tests/test_boolalg.py
+++ b/sympy/logic/tests/test_boolalg.py
@@ -737,6 +737,11 @@ def test_canonical_atoms():
     assert false.canonical == false
 
 
+def test_negated_atoms():
+    assert true.negated == false
+    assert false.negated == true
+
+
 def test_issue_8777():
     assert And(x > 2, x < oo).as_set() == Interval(2, oo, left_open=True)
     assert And(x >= 1, x < oo).as_set() == Interval(1, oo)