<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234". See
https://github.com/blog/1506-closing-issues-via-pull-requests . Please also
write a comment on that issue linking back to this pull request once it is
open. -->
Fixes #15869 

#### Brief description of what is fixed or changed
I have added a property `negated` to `Relational`s that returns the opposite Relational, so `Eq(x, y).negated` gives `Ne(x, y)`. This has only minor benefits over ~, one is that it ignores evaluate = False, so code that relies on negating a relation, primarily in `logic`, see #15869. The purpose of this code is to compare temporary expressions, so not to actually evaluate the expressions, so from that perspective it should be OK, just as the property `canonical` returns a slightly different expression, independent of evaluate. Hence, it should be used with a bit of care.

I also added the property to `BooleanAtom` which makes sense because the result of a `Relational` is a `BooleanAtom`, but doesn't make sense as no other objects in `logic` has that property. From a coding perspective it is only used in a few places (at the moment, only checked quite few cases, could be more).

#### Other comments
Not sure if there should be a release note. I guess it can be convenient for the user to know about it sometimes.

~~I also believe that it is slightly faster than using Not as there should be more processing steps involved with Not. Not a major factor, but no drawback either.~~ Timeit tests show that this is < 10% slower. Not obvious why.

Finally, I split the tests for #8444 as not all of them fails at the moment. Better to have those working is an executed test. Slightly unrelated, but better put it in here as I happened to see it now.

Oh, yeah, I was thinking quite long before deciding on `negated`. One may imagine `inverted` or `complemented` as well. And probably a few more options. I know how to change it...

#### Release Notes

<!-- Write the release notes for this release below. See
https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more information
on how to write release notes. The bot will check your release notes
automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* core
   * Relationals now have a property negated, which returns the opposite to the relation, similar to ~.
<!-- END RELEASE NOTES -->