@smichr 2019-02-04
> deciding on `negated`

it could be `foo.not`