@asmeurer 2019-02-04
I'm not sure I completely follow the reasoning behind it, but I think it makes sense to have the property. Plus it will be useful if we ever decide to make `Not(Inequality)` be unevaluated.