@oscargus 2019-02-05
@asmeurer The reasoning is more or less related to what you point out. Some methods in logic relies on comparisons of the type if `(~c1).canonical == c2.canonical`. With `evaluate=False`, `(~c1)` returns a `Not`object and not a `Relational`. The `Not` object does not have a `canonical`property so the code fails with an error.

Related: I was surprised that the new code is actually slower than using ~ (although only in the statistical sense, it varies, and the variance is larger than the difference). However, as we know that the original object is a `Relational`, so the arguments should be "good to go" and that no evaluation should be needed (even if we do not explicitly state it), is there another, more efficient, way to create the new `Relational`?

The corresponding lines in `Not` (in evaluate so the object has been created etc):
 https://github.com/sympy/sympy/blob/1bd5d04ffe2b47b91fd03a868ca31c3c3f9e243e/sympy/logic/boolalg.py#L726-L737