@oscargus 2019-02-05
> > deciding on `negated`
> 
> it could be `foo.not`

The reason for using a verb is e.g. `reversed`. It sounded more consistent with that and `canonical` (although I realize that it actually should be called `canonicalized` or something...). Also, `is_*`.

It seems like the Python name is actually `invert` so I think my new suggestion is `inverted`. 

(Now, being a non-native English speaker, my feeling may be wrong, but it was primarily for language consistency reasons.)