@asmeurer 2019-02-05
As far as the name, negated seems good to me. Invert doesn't exactly mean logical negation, and not is a keyword so you can't use it as a method name. 

As for reevaluating the arguments, you should be able to use evaluate=False. There may also be a class method constructor, I'm not sure. There is one on Mul, (_from_args or something like that), but I don't know about other objects.