@oscargus 2019-02-05
I improved the speed slightly by calling the `__new__` method of `Relational` directly. This decreased the time from 70 to 45 µs (for `Eq`and `Ne`, the other seems faster, now about 15 µs). For `reversed` it reduced from 50 to 45 µs  (again, for `Eq`and `Ne`). I do not fully understand why `negated` was slower than `reversed`earlier, but either way some improvement.

I also added the lambda version of a possibly new `Relational` using ~ as a comment. This will probably help a bit anyway in the unlikely case of a new `Relational`.

Finally, I added some notes trying to outline the difference to `~`/`Not`.