@oscarbenjamin 2019-02-05
Timings in SymPy are difficult to measure because of caching. Repeating the same measurement gives different results and it's not always clear whether the warm or cold cache time is the one you should worry about it in different situations:

```julia
In [1]: %time s=dsolve(f(x).diff(x)+y**2*x, f(x))                                                                                              
CPU times: user 239 ms, sys: 3.54 ms, total: 243 ms
Wall time: 241 ms

In [2]: %time s=dsolve(f(x).diff(x)+y**2*x, f(x))                                                                                              
CPU times: user 171 ms, sys: 2.41 ms, total: 173 ms
Wall time: 172 ms

In [3]: %time s=dsolve(f(x).diff(x)+y**2*x, f(x))                                                                                              
CPU times: user 170 ms, sys: 2.28 ms, total: 172 ms
Wall time: 171 ms
```

I don't actually know how the cache fits into everything though. Is it possible that a cache is involved in your timings?

When the cache is involved a fair comparison always uses a new process and runs the timings with both cold and warm cache. Of course you don't know whether your cache got warmed up during the import of SymPy which takes a substantial amount of processing:
```console
(venv) enojb@IT035770:~/current/sympy/sympy$ time python -c 'import sympy'

real	0m0.669s
user	0m0.600s
sys	0m0.057s
(venv) enojb@IT035770:~/current/sympy/sympy$ time python -c 'import os'

real	0m0.050s
user	0m0.031s
sys	0m0.012s
```