@oscargus 2019-02-05
It is also important to test the right thing...

I did `%timeit Eq(x, y).negated`etc. But that also includes the creation of the `Eq` object, which is slower than `Ge` etc.

I cannot really be bothered to run an exhaustive test again, but the actual call to `negated`, assuming an already created object, takes less than 2 µs now. And as the total time, including creating the object has decreased quite a bit more than that, I guess that the relative savings for the change is rather significant.