@smichr 2019-02-05
`not.negated`